[#5] Fix problem when saving dashboards with filters
As we changed the HTML element of the filters' dropdown to a <select> and not a
<div> in 7903f3, and it was inside the dashboard edit form, we ended up sending
their values together when editing the dashboard.

To fix it, I had to move it to outside our <form>. This was a bit hard, as I
had to detect when we're in the dashboard edit form to enable
dragging/resizing/removing widgets from other ways.