[#5] Remove duplicated allowed filters
As we're allowing the user to create as many filters as she wants now, there's
no reason to create a dashboard with multiple allowed filters to begin with. If
the admin creates one, we remove every repeated allowed filter, leaving just
the first occurrence.

Ideally, this would also work in JavaScript. I mean, if I have a filter for
"date" already created, when I add another filter, "date" doesn't appear on my
list of possible filters. Unfortunately, this would require a bit more work,
so, for now, this solutions is good enough.