Add configurable defaults sizex and sizey values
Individual views can't configure them anymore (removed in CKAN's commit
fd71934). We used to default to `sizex = sizey = 2`. Now we default to
`sizex = 4`, `sizey = 3`, but allow that to be configured by changing
dashboard-view-edit's `widgetSizeX` and `widgetSizeY` options.

This is very unlikely, but was as easy as adding regular variables, and avoids
using magic numbers.